Move common seismic code into a separate file
As it has been pointed out, we start having too much duplication between
observations and config, yet we do not have a good place to put it into.
So creating a new file to keep common logic.

ReservoirData was chosen as a general term to describe
RFT/Seismic/Summary response or observation data.
