+++ src/ert/config/seismic_config.py
@@ -2,12 +2,12 @@ from __future__ import annotations
 
 import logging
 from pathlib import Path
-from typing import Any, ClassVar, Literal, Self
+from typing import Any, Literal, Self
 
 import numpy as np
 import polars as pl
-import scipy as sp
 
+from ert.config._reservoir_data_utils import SeismicData
 from ert.substitutions import substitute_runpath_name
 
 from .parsing import (
@@ -30,8 +30,6 @@ class SeismicConfig(SimulationResponseConfig):
     name: str = "seismic"
     type: Literal["seismic"] = "seismic"
 
-    TOLERANCE: ClassVar[float] = 0.1
-
     @property
     def expected_input_files(self) -> list[str]:
         return self.input_files
@@ -50,8 +48,7 @@ class SeismicConfig(SimulationResponseConfig):
         easts = df["east"].to_numpy()
         norths = df["north"].to_numpy()
         coordinates = np.column_stack([easts, norths])
-        tree = sp.spatial.KDTree(coordinates)
-        too_close_pairs = tree.query_pairs(r=SeismicConfig.TOLERANCE * 2)
+        too_close_pairs = SeismicData.get_too_close_coordinate_pairs(coordinates)
         if too_close_pairs:
             too_close_coords = [
                 (
@@ -64,7 +61,7 @@ class SeismicConfig(SimulationResponseConfig):
                 "Seismic response coordinates with approximate locations "
                 f"{too_close_coords} fall inside of a tolerance radius. All seismic "
                 "response coordinates are expected to be more than "
-                f"{SeismicConfig.TOLERANCE * 2} m apart."
+                f"{SeismicData.TOLERANCE * 2} m apart."
             )
 
     def read_from_file(self, run_path: str, iens: int, iter_: int) -> pl.DataFrame:
@@ -108,57 +105,4 @@ class SeismicConfig(SimulationResponseConfig):
             name="seismic",
             input_files=files,
             keys=[Path(f).stem for f in files],
-        )
-
-    @classmethod
-    def use_observation_locations_in_respective_responses(
-        cls, responses: pl.DataFrame, observations: pl.DataFrame
-    ) -> pl.DataFrame:
-        """Unify response and observation locations.
-
-        Replace the east and north coordinates in the response dataframe with the
-        corresponding coordinates from the observation dataframe, if they are within the
-        tolerance radius. Drop response otherwise. Required to match responses to
-        observations regardless of location precision discrepancies.
-        """
-        candidates = responses.rename(
-            {
-                "east": "east_res",
-                "north": "north_res",
-                "response_key": "response_key_res",
-            }
-        ).join_where(
-            observations.rename(
-                {
-                    "east": "east_obs",
-                    "north": "north_obs",
-                    "response_key": "response_key_obs",
-                }
-            ),
-            pl.col("response_key_res") == pl.col("response_key_obs"),
-            pl.col("east_obs") >= pl.col("east_res") - cls.TOLERANCE,
-            pl.col("east_obs") <= pl.col("east_res") + cls.TOLERANCE,
-            pl.col("north_obs") >= pl.col("north_res") - cls.TOLERANCE,
-            pl.col("north_obs") <= pl.col("north_res") + cls.TOLERANCE,
-        )
-        matched_on_location = (
-            candidates.filter(
-                (pl.col("east_obs") - pl.col("east_res")).pow(2)
-                + (pl.col("north_obs") - pl.col("north_res")).pow(2)
-                <= cls.TOLERANCE**2
-            )
-        ).select(["response_key_res", "east_res", "north_res", "east_obs", "north_obs"])
-
-        return (
-            responses.join(
-                matched_on_location,
-                left_on=["response_key", "east", "north"],
-                right_on=["response_key_res", "east_res", "north_res"],
-                how="inner",
-            )
-            .with_columns(
-                east=pl.col("east_obs"),
-                north=pl.col("north_obs"),
-            )
-            .drop(["east_obs", "north_obs"])
         )
